Safari-609.1.11.1/LayoutTests/svg/animations/svgenum-animation-11-expected.txt

SVG 1.1 dynamic animation tests

Test BlendModeType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS feBlend.mode.animVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS feBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS feBlend.mode.animVal is SVGFEBlendElement.SVG_FEBLEND_MODE_NORMAL
PASS feBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS feBlend.mode.animVal is SVGFEBlendElement.SVG_FEBLEND_MODE_NORMAL
PASS feBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS feBlend.mode.animVal is SVGFEBlendElement.SVG_FEBLEND_MODE_MULTIPLY
PASS feBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS feBlend.mode.animVal is SVGFEBlendElement.SVG_FEBLEND_MODE_MULTIPLY
PASS feBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS feBlend.mode.animVal is SVGFEBlendElement.SVG_FEBLEND_MODE_SCREEN
PASS feBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS feBlend.mode.animVal is SVGFEBlendElement.SVG_FEBLEND_MODE_SCREEN
PASS feBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS feBlend.mode.animVal is SVGFEBlendElement.SVG_FEBLEND_MODE_DARKEN
PASS feBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS feBlend.mode.animVal is SVGFEBlendElement.SVG_FEBLEND_MODE_DARKEN
PASS feBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS feBlend.mode.animVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS feBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS successfullyParsed is true

TEST COMPLETE
